Clean up code a bit more
