+++ optimal_adp/regret.py
@@ -202,105 +202,6 @@ def rescale_adp_to_picks(updated_adp: dict[str, float]) -> dict[str, float]:
     return rescaled_adp
 
 
-def validate_position_hierarchy(
-    updated_adp: dict[str, float], players: list[Player]
-) -> bool:
-    """Validate that position hierarchy is maintained after ADP update.
-
-    Within each position, players with higher AVG should have lower (earlier) ADP.
-
-    Args:
-        updated_adp: Updated ADP values to validate
-        players: List of all players with their stats
-
-    Returns:
-        True if hierarchy is maintained, False if violated
-    """
-    # Group players by position
-    by_position: dict[str, list[Player]] = {}
-    for player in players:
-        if player.name in updated_adp:
-            if player.position not in by_position:
-                by_position[player.position] = []
-            by_position[player.position].append(player)
-
-    # Check hierarchy within each position
-    violations = 0
-    for position, position_players in by_position.items():
-        # Sort by ADP (earlier picks first)
-        sorted_by_adp = sorted(position_players, key=lambda p: updated_adp[p.name])
-
-        # Check that AVG scores are in descending order
-        prev_avg = float("inf")
-        for player in sorted_by_adp:
-            if player.avg > prev_avg:
-                logger.warning(
-                    f"Position hierarchy violation: {player.name} "
-                    f"(AVG: {player.avg}) has earlier ADP than higher-scoring player"
-                )
-                violations += 1
-            prev_avg = player.avg
-
-    is_valid = violations == 0
-    if is_valid:
-        logger.debug("Position hierarchy validation passed")
-    else:
-        logger.warning(f"Position hierarchy violations: {violations}")
-
-    return is_valid
-
-
-def validate_position_hierarchy_detailed(
-    updated_adp: dict[str, float], players: list[Player]
-) -> tuple[bool, list[str]]:
-    """Validate position hierarchy and return detailed violation list.
-
-    Within each position, players with higher AVG should have lower (earlier) ADP.
-
-    Args:
-        updated_adp: Updated ADP values to validate
-        players: List of all players with their stats
-
-    Returns:
-        Tuple of (is_valid, list_of_violations)
-    """
-    # Group players by position
-    by_position: dict[str, list[Player]] = {}
-    for player in players:
-        if player.name in updated_adp:
-            if player.position not in by_position:
-                by_position[player.position] = []
-            by_position[player.position].append(player)
-
-    # Check hierarchy within each position
-    violations: list[str] = []
-    for position, position_players in by_position.items():
-        # Sort by ADP (earlier picks first)
-        sorted_by_adp = sorted(position_players, key=lambda p: updated_adp[p.name])
-
-        # Check that AVG scores are in descending order
-        prev_player: Player | None = None
-        for player in sorted_by_adp:
-            if prev_player is not None and player.avg > prev_player.avg:
-                violation_msg = (
-                    f"{position}: {prev_player.name} (AVG: {prev_player.avg:.1f}, "
-                    f"ADP: {updated_adp[prev_player.name]:.1f}) ranked before "
-                    f"{player.name} (AVG: {player.avg:.1f}, "
-                    f"ADP: {updated_adp[player.name]:.1f})"
-                )
-                violations.append(violation_msg)
-                logger.warning(f"Position hierarchy violation: {violation_msg}")
-            prev_player = player
-
-    is_valid = len(violations) == 0
-    if is_valid:
-        logger.debug("Position hierarchy validation passed")
-    else:
-        logger.warning(f"Position hierarchy violations: {len(violations)}")
-
-    return is_valid, violations
-
-
 def check_convergence(
     initial_adp: dict[str, float], final_adp: dict[str, float]
 ) -> int:
